reworking on aligning squares
to calibrate hover mouse over square ( can now be in your farm).

This no longer checks for the purple outline when you hover over a square.

It now checks the inside of the square and checks for a colour, if it isn't that colour its position gets added to array. For this we are checking the yellow boarder around our plants. For testing this draws the added positions on the screen so you can see that it works

Soon I'll be added the same but so the outline can be drawn, It'll be added to the square information for the bot so each clickable square can be properly placed in the bots info on where each farm square is