new solution for positing
the new solution is. Every time you hover over a square, the top purple corner is displayed through any plants or object. with this I can position the square precisely on every square it moves to. Thank you devs for adding this small thing.

Right now, all it is, is moving the mouse to the top corner, but it works and I can now move it forward.