updated console ingame pos image detection
update this so its way more consistent.

Now pytesseract looks at the console and finds the last word with 'Tile'

from there it checks the values next to it

clean up values

convert them to ints

works.

need to add fix where if the amount of clicks on a tile is 10 or more the positioning of checking is off and causes an error